+++ HTTPLayer/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6]
+### Added
+- Implemented change Header Value on the Config Interface
+- Implemented remove Header on the Config Interface
+
 ## [0.0.5]
 ### Modified
 - Fix Generics and Object Encoding problems